Use list instead of generator for messages
We were using a generator expression for `messages`. The problem was that we
were logging the `messages` first, which had the side-effect of leaving the
generator in a consumed state.

This meant that when we later tried to send the messages, there would be
nothing left to send.

The quick solution is to just use a list where a new iterator over it will be
created with each `for` expression.